New sample rates
Now the sample rate should be a multiple of 3KHz from 3KHz to 192KHz.
higher frequencies should work but are locked because they have never been tested.
For the rates (strictly) lower than 48KHz, a restriction applies
as the output wave-form should be the continuous -K C only

The low rates have been designed to save on file or display the modulation envelope
while reducing the data size (per 16). They should never be sent to a sound-card.