IO parameters improvement and main_loop less dependant from jackaudio
The bundle between the samples generated by the software and the ones to the output
is more generic. It is no longer assumed jackaudio is float.
The output modules sends the hash code of the type and the software look for a match.
BY this way, no specific include files are needed in main_loop anymore.

That is the first step to make jackaudio optional and get an application
to convert formats or generate files only. Indeed, it sends a hash code
of jack_default_audio_sample_t which is an alias on float.

The IO parameters are now more object and include the output sound file management
-> side effect: -o xyz -f ztu -F text does not work anymore.
-> it should be replaced by -o xyz -F text -f ztu.

The output file format has now char and long as output format.
The default is short. Others are -F 8 or -F char, -F32 or -F long.

Some more cleanup is needed. Removing some commented out code,
  move the type supported, unsupported and planned into a single base function