Slew rate range internal change
The slew rate has now a better resolution on long raises and falls.
The text output parameters tell the previous and next possible value.
  Since we work with a value that is the derivative,
  talking about a constant step is a non-sense.
That moves the high limit to 2.666mS,
  means every parameter higher is constrained to this value.
  A bug has been fixed on that point.

There is no negative impact with the mnemonics.
The VHDL has to be updated to balance.
The midi mapped codes has to be modified, as now the SR is 2 times slower.